타인에게 알람을 설정하고 관리할 수 있게 변경함

+++ app/src/main/java/org/siwonlee/alarmapp12/Alarm_Data.kt
@@ -31,7 +31,7 @@ fun Alarm_Data(map: HashMap<String, Any>): Alarm_Data {
     var phr: Int? = (map["phr"] as Long).toInt()
     var pmin: Int? = (map["pmin"] as Long).toInt()
     var qr: String? = map["qr"] as String?
-    var solver: Int? = map["solver"] as Int
+    var solver: Int? = (map["solver"] as Long).toInt()
     var sound: String? = map["sound"] as String
     var switch: MutableList<Boolean> = map["switch"] as MutableList<Boolean>
     var timeInMillis: Long? = map["timeInMillis"] as Long
